Generate workflow templates

Generates workflow templates for systemPipeR package. The template environemts contain a predifined directory structure along with run parameter files and sample data. The strucutre of the workflow templates and the sample data are described in all details in the Overview Vignette of the systemPipeR package.

genWorkenvir(workflow, mydirname=NULL, bam=FALSE, ref="master", subdir=NULL, url=NULL, urlname=NULL)

Arguments

workflow

character string of workflow templates to be generated. Supported values can be checked with the \link{availableWF}() function. Workflow name containing a '/' are treated as GitHub repositories and installed using BiocManager::\link{install_github}(). For Github workflow templates, the format systemPipeR/repo is required. See Details.

mydirname

Specifies the name of the workflow directory. The default NULL uses the name of the chosen workflow. An error is issued if a directory of the same name and path exists already.

bam

If bam=TRUE pregenerated short read alignment (BAM) files will be included in the results directory of the workflow environment. Note, these BAM files have been generated with the HISAT2 aligner using the FASTQ files provided in the data directory. The default bam=FALSE omits this step meaning no BAM files will be copied into the results directory.

ref

Desired GitHub reference for the branch name. Default to master branch.

subdir

subdirectory within GitHub repo that contains the R package, if it is required.

url

character string of a specifies the URL of a different version of the R Markdown workflow template or other file to download. The default NULL copies the current version available in the `systemPipeRdata` or the workflow template on GitHub being selected.

urlname

character string with the name where the downloaded file is saved. This is argument is required when the url is provided.

Value

Workflow directory containing sample data and parameter files along with the following subdirectories:

param/

stores parameter files

data/

stores input data

results/

stores output results

For more details, please consult the Overview Vignette (HTML) of the systemPipeR package (http://bioconductor.org/packages/systemPipeR).

Details

When installing GitHub Workflow Packages under systemPipeR Organization, it is requested to the remotes package function BiocManager::\link{install_github}() to build vignettes and also install all the dependencies, via dependencies=TRUE, build_vignettes=TRUE.

Check the output of \link{availableWF}() to the current workflow templates available on systemPipeR Organization. The argument

For an interactive() session, the readline() function provides the option choose between proceeding or not, through options: yes or no. For non-interactive use, if there is no package install, the option yes will be selected.

Author

Thomas Girke and Daniela Cassol

Examples

## Return location of sample data
samplepaths <- pathList()
if (FALSE) {
## Generate varseq workflow environment
genWorkenvir(workflow="varseq", mydirname=NULL, url=NULL, urlname=NULL)
setwd("varseq")

## List Workflow Templates from \code{systemPipeRdata} package and \code{systemPipeR} Organization 
availableWF(github = TRUE)
## Generate 'systemPipeR/systemPipeChIPseq' workflow environment 
genWorkenvir(workflow="systemPipeR/systemPipeChIPseq", mydirname=NULL, ref="master", subdir=NULL)
setwd("systemPipeChIPseq")

## Download a specific R Markdown file
genWorkenvir(workflow="systemPipeR/systemPipeRNAseq", mydirname="rnaseq", url = "https://raw.githubusercontent.com/systemPipeR/systemPipeRNAseq/cluster/vignettes/systemPipeRNAseq.Rmd", urlname = "rnaseq_V-cluster.Rmd")
}