Subsetting fastq data

Returns subsets of fastq files data based on specific mapping regions or list of genes or GRanges object.

getSubsetReads(args,
  geneList = NULL,
  gr = NULL,
  MappingRegion = 1:1e+05,
  sample_range = 90000:1e+05,
  truncate_refs = TRUE,
  id_read_number = TRUE,
  annotation = "data/tair10.gff",
  reference = "data/tair10.fasta",
  annot_outname = "tair10_sub.gff",
  ref_outname = "tair10_sub.fasta",
  outdir = "data/subset/",
  silent = FALSE
)

Arguments

args

object of class SYSargs2.

geneList

selected genes list to retrieve the reads from the fastq file.

gr

an object containing genomic ranges to retrieve the reads from the fastq file.

MappingRegion

integers ranges of start and end of chromosome position to retrieve the reads from the fastq file.

sample_range

random range to subsetted the fastq file.

truncate_refs

logical. If TRUE it will generate reference genome and annotation subset file.

id_read_number

if fastq file contains sequence name with read number ($ri - --defline-seq '@$sn[_$rn]/$ri').

annotation

path to annotation file.

reference

path to reference genome.

annot_outname

character name of the annotation output file.

ref_outname

character name of the reference genome output file.

outdir

path to output directory.

silent

if set to TRUE, all messages returned by the function will be suppressed.

Value

Workflow directory containing sample data and parameter files along with the following subdirectories:

param/

stores parameter files

data/

stores input data

results/

stores output results

For more details, please consult the Overview Vignette (HTML) of the systemPipeR package (http://bioconductor.org/packages/systemPipeR).

Author

Thomas Girke, Shiyuan Guo and Daniela Cassol

Examples

if (FALSE) {
getSubsetReads(args, MappingRegion = 1:900, sample_range = 800:900, outdir = "data/subset/", silent = FALSE)
getSubsetReads(args, MappingRegion = 1:900, sample_range = NULL, outdir = "data/subset/", silent = FALSE)
}